requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiofiles==0.7.0
    # via -r requirements.in
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via uvicorn
fastapi==0.65.1
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
idna==2.10
    # via requests
jinja2==3.0.1
    # via -r requirements.in
markupsafe==2.0.1
    # via jinja2
pydantic==1.8.2
    # via fastapi
python-dotenv==0.17.1
    # via -r requirements.in
requests==2.25.1
    # via spotipy
six==1.16.0
    # via spotipy
spotipy==2.18.0
    # via -r requirements.in
starlette==0.14.2
    # via fastapi
typing-extensions==3.10.0.0
    # via pydantic
urllib3==1.26.4
    # via
    #   requests
    #   spotipy
uvicorn==0.13.4
    # via -r requirements.in
